+++ README.md
@@ -18,16 +18,19 @@
 
 ## 新特性:
 
-## 1.1.0 由于新接口更加便利 这里不对1.0.0版本接口做完全兼容
+## 1.2.0 预更新
+- tracker http 负载均衡
 
+## 1.1.0
+备注：由于新接口更简单便利 这里不对1.0.0版本接口做兼容
 - FastDfs聚合tracker和http功能
-- 重构分片上传接口
+- 重构分片上传
 - 优化删除文件接口 文件不存在则抛出异常
 - 其他细节优化和bug处理
 
 ## 1.0.0
 
-- 升级fastdfs为6.07最新版本
+- 升级fastdfs 6.07 version
 - 大文件分片上传
 - token防盗链功能
 - 新增regenerateAppenderFile接口 支持appender文件改为普通文件 仅支持6.0.2以上版本
