+++ src/main/java/com/ykrenz/fastdfs/multipart/MultipartAttachmentAccessor.java
@@ -1,5 +1,8 @@
-package com.ykrenz.fastdfs.cache;
+package com.ykrenz.fastdfs.multipart;
 
+/**
+ * 分片信息访问器
+ */
 public interface MultipartAttachmentAccessor {
 
     void put(String groupName, String path, MultipartUploadAttachment attachment);
